// src/pages/items/components/ItemsDrawer/tabs/InformationTab/__generated__/InformationTabInfoSectionFragment.graphql.ts
/**
 * @generated SignedSource<<00d831e249238a6a78d51bc5c33a2bd2>>
 * @lightSyntaxTransform
 * @nogrep
 */

/* tslint:disable */
/* eslint-disable */
// @ts-nocheck

import { Fragment, ReaderFragment } from 'relay-runtime';
import { FragmentRefs } from "relay-runtime";
export type InformationTabInfoSectionFragment$data = {
  readonly barcode: string;
  readonly cost: number;
  readonly creationDate: any;
  readonly currentStock: number;
  readonly id: string;
  readonly isActive: boolean;
  readonly isService: boolean;
  readonly markup: number;
  readonly name: string;
  readonly price: number | null;
  readonly sku: string;
  readonly " $fragmentType": "InformationTabInfoSectionFragment";
};
export type InformationTabInfoSectionFragment$key = {
  readonly " $data"?: InformationTabInfoSectionFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"InformationTabInfoSectionFragment">;
};

const node: ReaderFragment = {
  "argumentDefinitions": [],
  "kind": "Fragment",
  "metadata": null,
  "name": "InformationTabInfoSectionFragment",
  "selections": [
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "id",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "name",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "barcode",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "cost",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "isActive",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "sku",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "price",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "markup",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "currentStock",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "creationDate",
      "storageKey": null
    },
    {
      "alias": null,
      "args": null,
      "kind": "ScalarField",
      "name": "isService",
      "storageKey": null
    }
  ],
  "type": "ItemType",
  "abstractKey": null
};

(node as any).hash = "da2238a44d57aae5e80514a3fffe2f76";

export default node;